+++ src/Code/ResourseHandle/Utility.hpp
@@ -23,6 +23,7 @@ void centerOrigin(sf::Text& text);
 float toDegree(float radian);
 float toRadian(float degree);
 
+void centerOrigin(sf::Text& text);
 void setScale(sf::Sprite& sprite, sf::IntRect rectangle);
 
 int randomInt(int exclusiveMax);